[tests] Move wait_for_services and test_admin_login to selenium
